feat: collapsed section peek for new arrivals
When entries arrive in a collapsed section, a peek slot opens
below the header showing the latest card with category glow.
Auto-retracts after 3s. Header gets +N badge and brief glow
pulse. Tapping the peeked card expands the section.

Extract SwipeableCard to its own file to stay under line limit.
